requirements.txt
streamlit==1.26.1
openai==1.0.0
python-dotenv==1.0.0
plotly==5.16.1
kaleido==0.3.1
reportlab==4.0.3
pandas==2.1.1
Pillow==10.1.0
